HBASE-24140 : Move CandidateGenerator and their implementors out of StochasticLoadBalancer (#1458)
